Split structures.py.
