+++ databricks/koala/selection.py
@@ -171,7 +171,7 @@ class SparkDataFrameLocator(object):
         df._metadata = self.df._metadata.copy(
             column_fields=df._metadata.column_fields[-len(columns):])
         if cols_sel is not None and isinstance(cols_sel, Column):
-            from .structures import _col
+            from .series import _col
             return _col(df)
         else:
             return df
